load more: fix my library refetch on close

+++ web/src/components/block/product/installModal.jsx
@@ -6,6 +6,7 @@ import { useSectionListContext } from '~/context';
 import { useCallback, useState } from 'react';
 import { useManage } from '~/talons/section/useManage';
 import ThemeList from './themeList';
+import { SECTION_TYPE_SIMPLE } from '~/constants/index.js';
 
 const useInstallModal = (refetch) => {
   const [{ activeSection, modal }, { setActiveSection, setModalLoading, setModal }] = useSectionListContext();
@@ -35,22 +36,25 @@ const InstallModal = props => {
   const talonManageProps = useManage({ section: section });
   const {
     dataUpdateLoading,
-    dataDeleteLoading
+    dataDeleteLoading,
+    handleDelete: uninstallSectionFromTheme,
+    currentThemeSelected: selectedTheme,
+    installed,
   } = talonManageProps;
   const [isShowConfirm, setIsShowConfirm] = useState(false);
-  const [confirmAction, setConfirmAction] = useState(() => {});
-  const [currentThemeSelected, setCurrentThemeSelected] = useState(undefined);
+  // const [confirmAction, setConfirmAction] = useState(() => {});
+  // const [currentThemeSelected, setCurrentThemeSelected] = useState(undefined);
 
   // Stop auto scroll if modal is active
   if (splide?.current?.splide && !!activeSection?.url_key) {
     splide.current.splide.Components.Autoplay.pause();
   }
 
-  const confirmDelete = useCallback(() => {
-    setConfirmAction(() => talonManageProps.handleDelete);
-    setIsShowConfirm(true);
-    setCurrentThemeSelected(talonManageProps.currentThemeSelected);
-  });
+  // const confirmDelete = () => {
+  //   // setConfirmAction(() => talonManageProps.handleDelete);
+  //   setIsShowConfirm(true);
+  //   // setCurrentThemeSelected(talonManageProps.currentThemeSelected);
+  // };
 
   const handleClose = useCallback(() => {
     handleCloseModal();
@@ -59,29 +63,31 @@ const InstallModal = props => {
     }
   }, []);
 
+  if (isShowConfirm) {
+    return <ModalConfirm section={section} theme={selectedTheme} isOpen={isShowConfirm} setIsOpen={setIsShowConfirm} onConfirm={uninstallSectionFromTheme} />
+  }
+
+  if (!show || !activeSection) return null;
   return (
-    !isShowConfirm
-    ? <Modal
+    <Modal
       open={show}
       onClose={handleClose}
       title={`Install "${activeSection?.name ?? 'section'}" to theme`}
       primaryAction={{
-        content: (!loadingWithoutData && talonManageProps.installed) ? 'Reinstall to theme' : 'Install to theme',
+        content: (!loadingWithoutData && installed) ? 'Reinstall to theme' : 'Install to theme',
         disabled: dataDeleteLoading || dataUpdateLoading || loadingWithoutData || !talonManageProps.options.length || !section?.actions?.install,
         loading: talonManageProps?.executeSection === activeSection?.url_key ? talonManageProps.dataUpdateLoading : false,
         onAction: talonManageProps.handleUpdate
       }}
       secondaryActions={
-        activeSection?.type_id == '1' // Simple section
-        ? [
-          {
+        activeSection && parseInt(activeSection.type_id) === SECTION_TYPE_SIMPLE ?
+          [{
             destructive: true,
             content: 'Delete from theme',
             disabled: dataDeleteLoading || dataUpdateLoading || loadingWithoutData || (section?.installed ? !talonManageProps.installed : true),
             loading: talonManageProps?.executeSection === activeSection?.url_key ? talonManageProps.dataDeleteLoading : false,
-            onAction: confirmDelete
-          },
-        ]
+            onAction: () => setIsShowConfirm(true),
+          }]
         : []
       }
     >
@@ -94,7 +100,6 @@ const InstallModal = props => {
         </BlockStack>
       </Modal.Section>
     </Modal>
-    : <ModalConfirm section={section} theme={currentThemeSelected} isOpen={isShowConfirm} setIsOpen={setIsShowConfirm} onConfirm={confirmAction} />
   );
 };
